Use AST for hbs extraction
We used to use xgettext-template to extract the texts from .hbs files. While this worked, it was very slow - as we needed to do a lot of file read & write operations.

With this change, we instead use the Glimmer VM to generate an AST of a .hbs file, and parse the translation strings out from there. This is considerably faster.