Ensure ember-l10n works with Prember
Prember will pre-render pages before the postBuild() hook is fired, leading to issues with the way we generate the manifest for FastBoot.
Instead, we now generate the asset map via a Broccoli Plugin (in postprocessTree), which works as expected in both regular FastBoot as well as in Prember.

We also need to make sure this addon runs before ember-cli-uglify, as otherwise the `postprocessTree` hooks might conflict.